Restructured files to make things more organized
