--- src/pages/AddEmployee.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const AddEmployee = () => {
-  return (
-    <div>AddEmployee</div>
-  )
-}
-
-export default AddEmployee;
